feat: Migrate all builders into one
