+++ src/lib.rs
@@ -13,11 +13,8 @@
 //!
 //! ## Payload
 //!
-//! Building the notification payload should be done with the corresponding builders:
-//!
-//! * [PlainNotificationBuilder](request/notification/struct.PlainNotificationBuilder.html) for text only messages.
-//! * [SilentNotificationBuilder](request/notification/struct.SilentNotificationBuilder.html) for silent notifications with custom data.
-//! * [LocalizedNotificationBuilder](request/notification/struct.LocalizedNotificationBuilder.html) for localized rich notifications.
+//! Building the notification payload should be done with the [DefaultNotificationBuilder](request/notification/struct.DefaultNotificationBuilder.html) for most use-cases.
+//! There is also the [WebNotificationBuilder](request/notification/struct.WebNotificationBuilder.html) in the case you need to send notifications to safari
 //!
 //! The payload generated by the builder [can hold a custom data
 //! section](request/payload/struct.Payload.html#method.add_custom_data),
@@ -34,11 +31,12 @@
 //! ## Example sending a plain notification using token authentication:
 //!
 //! ```no_run
-//! # use a2::{PlainNotificationBuilder, NotificationBuilder, Client, Endpoint};
+//! # use a2::{DefaultNotificationBuilder, NotificationBuilder, Client, Endpoint};
 //! # use std::fs::File;
 //! # #[tokio::main]
 //! # async fn main() -> Result<(), Box<dyn std::error::Error + Send + Sync>> {
-//! let mut builder = PlainNotificationBuilder::new("Hi there");
+//! let mut builder = DefaultNotificationBuilder::new();
+//! builder.set_body("Hi there");
 //! builder.set_badge(420);
 //! builder.set_category("cat1");
 //! builder.set_sound("ping.flac");
@@ -64,7 +62,7 @@
 //! #[macro_use] extern crate serde;
 //!
 //! use a2::{
-//!     Client, Endpoint, SilentNotificationBuilder, NotificationBuilder, NotificationOptions,
+//!     Client, Endpoint, DefaultNotificationBuilder, NotificationBuilder, NotificationOptions,
 //!     Priority,
 //! };
 //! use std::fs::File;
@@ -82,7 +80,8 @@
 //!         is_paying_user: false,
 //!     };
 //!
-//!     let mut payload = SilentNotificationBuilder::new()
+//!     let mut payload = DefaultNotificationBuilder::new()
+//!         .set_content_available()
 //!         .build("device-token-from-the-user",
 //!         NotificationOptions {
 //!             apns_priority: Some(Priority::Normal),
@@ -119,8 +118,8 @@ pub mod response;
 mod signer;
 
 pub use crate::request::notification::{
-    CollapseId, LocalizedNotificationBuilder, NotificationBuilder, NotificationOptions, PlainNotificationBuilder,
-    Priority, SilentNotificationBuilder, WebNotificationBuilder, WebPushAlert,
+    CollapseId, DefaultNotificationBuilder, NotificationBuilder, NotificationOptions, Priority, WebNotificationBuilder,
+    WebPushAlert,
 };
 
 pub use crate::response::{ErrorBody, ErrorReason, Response};
